move getContent() to util file to reuse in other components, add tsconfig file, add wordpress API calls to populate top level tab content
