+++ pages/transfer.tsx
@@ -4,8 +4,11 @@ import Row from "react-bootstrap/Row";
 import Breadcrumb from "react-bootstrap/Breadcrumb";
 
 import HtanNavbar from "../components/HtanNavbar";
+import {getContent} from "../util";
 
-const Transfer = () => (
+const Transfer = () => {
+    const content = getContent("data-transfer","summary-blurb");
+    return (
     <>
         <HtanNavbar/>
         <Container>
@@ -21,19 +24,11 @@ const Transfer = () => (
                 <h1>Data Transfer</h1>
             </Row>
             <Row className="mt-3">
-                <p>
-                    Lorem ipsum dolor sit amet, consectetur adipiscing elit, sed do eiusmod tempor incididunt ut
-                    labore et dolore magna aliqua. Ut enim ad minim veniam, quis nostrud exercitation ullamco
-                    laboris
-                    nisi ut aliquip ex ea commodo consequat. Duis aute irure dolor in reprehenderit in voluptate
-                    velit
-                    esse cillum dolore eu fugiat nulla pariatur. Excepteur sint occaecat cupidatat non proident,
-                    sunt
-                    in culpa qui officia deserunt mollit anim id est laborum.
-                </p>
+                <span dangerouslySetInnerHTML={{__html: content}} />
             </Row>
         </Container>
     </>
-);
+    )
+};
 
 export default Transfer
